feature: combine view mode with edit mode
