+++ backend/templates.json
@@ -576,5 +576,27 @@
     "usageCount": 0,
     "tags": [],
     "points": 15
+  },
+  {
+    "title": "asdfas",
+    "criteria": [],
+    "key": "d980afee-c23f-42de-a630-4c65cfd86ca2",
+    "description": "",
+    "createdAt": "2025-02-05T02:58:21.326Z",
+    "lastUsed": null,
+    "usageCount": 0,
+    "tags": [],
+    "points": 0
+  },
+  {
+    "title": "asdfasdf",
+    "criteria": [],
+    "key": "14497d83-6986-43ca-bbd0-2fe7ac8ed002",
+    "description": "",
+    "createdAt": "2025-02-05T02:59:09.025Z",
+    "lastUsed": null,
+    "usageCount": 0,
+    "tags": [],
+    "points": 0
   }
 ]
